Appointment corrections
values of t array starting from one for convolve and msd.
Rectified fit on smaller scale.
Corrected autocorelation with larger number of points
added single traj for coefficient determination.
Problematic with large value of spatial frequences (not hz). maybe need to lower the freq  and dt and increase the number of points